+++ src/main/kotlin/com/collect/invest/dao/WalletsDao.kt
@@ -5,8 +5,8 @@ import com.collect.invest.dao.entity.WalletsEntity
 interface WalletsDao {
     fun createWallet(id: Long)
     fun getById(userId: Long): WalletsEntity?
-    fun topupBalance(userId: Long, amount: Int)
-    fun withdrawBalance( userId: Long, amount: Int)
+    fun topupBalance(userId: Long, amount: Double)
+    fun withdrawBalance( userId: Long, amount: Double)
     fun updateStatus(userId: Long, status: String)
 
 }
